# backend/src/main/java/com/sajuppugi/fortune/reading/api/SuneungReadingController.java
package com.sajuppugi.fortune.reading.api;

import com.sajuppugi.common.api.ApiException;
import com.sajuppugi.common.api.ApiResponse;
import com.sajuppugi.common.api.ErrorCode;
import com.sajuppugi.common.idempotency.IdempotencyKey;
import com.sajuppugi.catalog.domain.PurchaseQuote;
import com.sajuppugi.catalog.application.QuoteFundingService;
import io.swagger.v3.oas.annotations.media.Schema;
import com.sajuppugi.fortune.generation.domain.GenerationModels.GeneratedSection;
import com.sajuppugi.fortune.generation.domain.GenerationModels.FortuneType;
import com.sajuppugi.fortune.reading.application.GeneralReadingService;
import com.sajuppugi.fortune.reading.application.ReadingQueryService;
import com.sajuppugi.fortune.reading.application.SuneungReadingService;
import com.sajuppugi.fortune.reading.application.SuneungReadingService.PurchaseResult;
import com.sajuppugi.fortune.reading.domain.OwnedReading;
import com.sajuppugi.fortune.reading.domain.OwnedReading.ProductOption;
import com.sajuppugi.fortune.reading.domain.RelationType;
import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import java.time.Instant;
import java.time.LocalDate;
import java.util.List;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1")
public class SuneungReadingController {
    private final SuneungReadingService service;
    private final GeneralReadingService generalService;
    private final ReadingQueryService queryService;
    private final QuoteFundingService funding;

    public SuneungReadingController(SuneungReadingService service, GeneralReadingService generalService,
                                     ReadingQueryService queryService, QuoteFundingService funding) {
        this.service = service;
        this.generalService = generalService;
        this.queryService = queryService;
        this.funding = funding;
    }

    @PostMapping("/reading-purchases")
    public ResponseEntity<ApiResponse<PurchaseResponse>> purchase(
            Authentication authentication,
            @RequestHeader("Idempotency-Key") String idempotencyKey,
            @Valid @RequestBody PurchaseRequest request) {
        PurchaseResult result;
        if (request.fortuneType() == null) {
            if (!request.isLegacySuneung()) throw new ApiException(ErrorCode.INVALID_REQUEST);
            result = service.purchase(userId(authentication), request.quoteId(), request.personId(),
                    new IdempotencyKey(idempotencyKey));
        } else {
            result = generalService.purchase(userId(authentication), request.quoteId(), request.fortuneType(),
                    request.productOption(), request.personId(), request.counterpartPersonId(),
                    request.relationType(), request.questionKey(), new IdempotencyKey(idempotencyKey));
        }
        HttpStatus status = result.status() == com.sajuppugi.fortune.reading.domain.ReadingPurchase.Status.FULFILLED
                ? HttpStatus.CREATED : HttpStatus.ACCEPTED;
        return ResponseEntity.status(status).body(ApiResponse.of(PurchaseResponse.from(result)));
    }

    @PostMapping("/quotes/fortune")
    public ResponseEntity<ApiResponse<SuneungQuoteResponse>> quote(
            Authentication authentication, @Valid @RequestBody QuoteRequest request) {
        if (request.fortuneType() == null && !request.isLegacySuneung()) {
            throw new ApiException(ErrorCode.INVALID_REQUEST);
        }
        UUID userId = userId(authentication);
        PurchaseQuote quote = request.fortuneType() == null
                ? service.issueQuote(userId, request.personId())
                : generalService.issueQuote(userId, request.fortuneType(), request.productOption(),
                        request.personId(), request.counterpartPersonId(), request.relationType(), request.questionKey());
        var funded = funding.get(userId, quote.id());
        Event event = request.fortuneType() == null
                ? new Event("CSAT", com.sajuppugi.fortune.reading.application.SuneungEventPolicy.EXAM_DATE) : null;
        return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.of(new SuneungQuoteResponse(
                quote.id(), quote.productSnapshot().code(), new Charged(
                        quote.productSnapshot().price().currency().name(), quote.productSnapshot().price().amount()),
                quote.expiresAt(), event, funded.walletBalance(), funded.balanceAfter(),
                funded.shortage(), funded.recommendedTopUp())));
    }

    @GetMapping("/readings")
    public ApiResponse<ReadingListResponse> list(
            Authentication authentication,
            @RequestParam(required = false) FortuneType fortuneType,
            @RequestParam(required = false) UUID personId,
            @RequestParam(required = false) String cursor,
            @RequestParam(required = false) Integer size) {
        ReadingQueryService.Page page = queryService.list(userId(authentication), fortuneType, personId, cursor, size);
        return ApiResponse.of(new ReadingListResponse(page.items().stream().map(ReadingSummary::from).toList(),
                page.nextCursor(), page.hasNext()));
    }

    @GetMapping("/readings/{readingId}")
    public ApiResponse<ReadingResponse> get(Authentication authentication, @PathVariable UUID readingId) {
        return ApiResponse.of(ReadingResponse.from(queryService.get(userId(authentication), readingId)));
    }

    private UUID userId(Authentication authentication) {
        try {
            return UUID.fromString(authentication.getName());
        } catch (RuntimeException exception) {
            throw new ApiException(ErrorCode.AUTHENTICATION_REQUIRED);
        }
    }

    public record PurchaseRequest(@NotNull UUID quoteId, @NotNull UUID personId,
                                  FortuneType fortuneType, ProductOption productOption, String questionKey,
                                  UUID counterpartPersonId, RelationType relationType) {
        boolean isLegacySuneung() {
            return productOption == null && questionKey == null && counterpartPersonId == null && relationType == null;
        }
    }

    public record QuoteRequest(@NotNull UUID personId, FortuneType fortuneType, ProductOption productOption,
                               String questionKey, UUID counterpartPersonId, RelationType relationType) {
        boolean isLegacySuneung() {
            return productOption == null && questionKey == null && counterpartPersonId == null && relationType == null;
        }
    }

    @Schema(requiredProperties = {"quoteId", "productCode", "charged", "expiresAt", "event", "walletBalance",
            "balanceAfter", "shortage", "recommendedTopUp"})
    public record SuneungQuoteResponse(UUID quoteId, String productCode, Charged charged,
                                Instant expiresAt, Event event, int walletBalance,
                                @Schema(types = {"integer", "null"}) Integer balanceAfter, int shortage,
                                @Schema(types = {"string", "null"}) String recommendedTopUp) {}

    public record PurchaseResponse(UUID purchaseId, UUID readingId, String status,
                                   String calculationVersion, String generationVersion, String contentVersion,
                                   Charged charged, Integer balance, boolean reused, ReadingResponse reading) {
        static PurchaseResponse from(PurchaseResult result) {
            return new PurchaseResponse(result.purchaseId(), result.readingId(), result.status().name(),
                    result.reading() == null ? null : result.reading().calculationVersion(),
                    result.reading() == null ? null : result.reading().generationVersion(),
                    result.reading() == null ? null : result.reading().contentVersion(),
                    result.charged() == null ? null : new Charged(
                            result.charged().currency().name(), result.charged().amount()),
                    result.balance(), result.reused(), result.reading() == null ? null : ReadingResponse.from(result.reading()));
        }
    }

    public record Charged(String currency, int amount) {}

    public record ReadingResponse(UUID id, String fortuneType, String productOption, String subjectDisplayName,
                                  Counterpart counterpart, String questionKey, Event event,
                                  List<GeneratedSection> sections, String calculationVersion,
                                  String generationVersion, String contentVersion, String generationMode, Instant createdAt,
                                  List<String> disclaimers, Talisman talisman) {
        static ReadingResponse from(OwnedReading reading) {
            return new ReadingResponse(reading.id(), reading.fortuneType().name(), reading.productOption().name(),
                    reading.subjectDisplayName(), reading.counterpartPersonId() == null ? null : new Counterpart(
                            reading.counterpartPersonId(), reading.counterpartDisplayName(), reading.relationType()),
                    reading.questionKey(), reading.eventDate() == null ? null : new Event("CSAT", reading.eventDate()),
                    reading.sections().sections(), reading.calculationVersion(), reading.generationVersion(),
                    reading.contentVersion(), reading.generationMode().name(), reading.createdAt(), List.of("FOR_ENTERTAINMENT"),
                    reading.talismanId() == null ? null : new Talisman(reading.talismanId(), reading.talismanStatus()));
        }
    }

    public record ReadingListResponse(List<ReadingSummary> items, String nextCursor, boolean hasNext) {}

    public record ReadingSummary(UUID id, String fortuneType, String productOption, String subjectDisplayName,
                                 String counterpartDisplayName, String questionKey, String generationMode,
                                 Instant createdAt) {
        static ReadingSummary from(OwnedReading reading) {
            return new ReadingSummary(reading.id(), reading.fortuneType().name(), reading.productOption().name(),
                    reading.subjectDisplayName(), reading.counterpartDisplayName(), reading.questionKey(),
                    reading.generationMode().name(), reading.createdAt());
        }
    }

    public record Counterpart(UUID personId, String displayName, String relationType) {}

    public record Event(String type, LocalDate date) {}

    public record Talisman(UUID id, String status) {}
}