# docs/openapi/api-v1.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponsePurchaseResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PurchaseResponse"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseReadingListResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ReadingListResponse"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseReadingResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ReadingResponse"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseSuneungQuoteResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/SuneungQuoteResponse"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Charged" : {
        "properties" : {
          "amount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "currency" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Counterpart" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "personId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "relationType" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Event" : {
        "properties" : {
          "date" : {
            "format" : "date",
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GeneratedSection" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "key" : {
            "enum" : [ "SUMMARY", "CURRENT_FLOW", "GOOD_PERIOD", "CAUTION", "ACTION_TIP", "RELATIONSHIPS", "STUDY_AND_WORK", "WEALTH_FLOW", "CONDITION", "LUCKY_POINT", "INCOME", "SPENDING_CAUTION", "MATCH_STRENGTH", "MATCH_CONFLICT", "COMMUNICATION", "RELATIONSHIP_TIP", "SPECIAL_STARS", "BALANCING_SPECIAL_STARS", "MISSING_ELEMENT", "EXAM_DAY", "EXAM_PERIODS", "FOCUS", "MEAL", "PREPARATION", "ANXIETY_MANAGEMENT" ],
            "type" : "string"
          },
          "sourceFactKeys" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "PurchaseRequest" : {
        "properties" : {
          "counterpartPersonId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "fortuneType" : {
            "enum" : [ "OVERALL", "LOVE", "WEALTH", "COMPATIBILITY", "SINSAL", "SUNEUNG" ],
            "type" : "string"
          },
          "personId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "productOption" : {
            "enum" : [ "READING_ONLY", "READING_WITH_TALISMAN" ],
            "type" : "string"
          },
          "questionKey" : {
            "type" : "string"
          },
          "quoteId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "relationType" : {
            "enum" : [ "FAMILY", "FRIEND", "LOVER", "CRUSH", "WORK_SCHOOL", "OTHER" ],
            "type" : "string"
          }
        },
        "required" : [ "personId", "quoteId" ],
        "type" : "object"
      },
      "PurchaseResponse" : {
        "properties" : {
          "balance" : {
            "format" : "int32",
            "type" : "integer"
          },
          "calculationVersion" : {
            "type" : "string"
          },
          "charged" : {
            "$ref" : "#/components/schemas/Charged"
          },
          "contentVersion" : {
            "type" : "string"
          },
          "generationVersion" : {
            "type" : "string"
          },
          "purchaseId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reading" : {
            "$ref" : "#/components/schemas/ReadingResponse"
          },
          "readingId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reused" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "QuoteRequest" : {
        "properties" : {
          "counterpartPersonId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "fortuneType" : {
            "enum" : [ "OVERALL", "LOVE", "WEALTH", "COMPATIBILITY", "SINSAL", "SUNEUNG" ],
            "type" : "string"
          },
          "personId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "productOption" : {
            "enum" : [ "READING_ONLY", "READING_WITH_TALISMAN" ],
            "type" : "string"
          },
          "questionKey" : {
            "type" : "string"
          },
          "relationType" : {
            "enum" : [ "FAMILY", "FRIEND", "LOVER", "CRUSH", "WORK_SCHOOL", "OTHER" ],
            "type" : "string"
          }
        },
        "required" : [ "personId" ],
        "type" : "object"
      },
      "ReadingListResponse" : {
        "properties" : {
          "hasNext" : {
            "type" : "boolean"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ReadingSummary"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReadingResponse" : {
        "properties" : {
          "calculationVersion" : {
            "type" : "string"
          },
          "contentVersion" : {
            "type" : "string"
          },
          "counterpart" : {
            "$ref" : "#/components/schemas/Counterpart"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "disclaimers" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "event" : {
            "$ref" : "#/components/schemas/Event"
          },
          "fortuneType" : {
            "type" : "string"
          },
          "generationMode" : {
            "type" : "string"
          },
          "generationVersion" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "productOption" : {
            "type" : "string"
          },
          "questionKey" : {
            "type" : "string"
          },
          "sections" : {
            "items" : {
              "$ref" : "#/components/schemas/GeneratedSection"
            },
            "type" : "array"
          },
          "subjectDisplayName" : {
            "type" : "string"
          },
          "talisman" : {
            "$ref" : "#/components/schemas/Talisman"
          }
        },
        "type" : "object"
      },
      "ReadingSummary" : {
        "properties" : {
          "counterpartDisplayName" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "fortuneType" : {
            "type" : "string"
          },
          "generationMode" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "productOption" : {
            "type" : "string"
          },
          "questionKey" : {
            "type" : "string"
          },
          "subjectDisplayName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SuneungQuoteResponse" : {
        "properties" : {
          "balanceAfter" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "charged" : {
            "$ref" : "#/components/schemas/Charged"
          },
          "event" : {
            "$ref" : "#/components/schemas/Event"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "productCode" : {
            "type" : "string"
          },
          "quoteId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "recommendedTopUp" : {
            "type" : [ "string", "null" ]
          },
          "shortage" : {
            "format" : "int32",
            "type" : "integer"
          },
          "walletBalance" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "balanceAfter", "charged", "event", "expiresAt", "productCode", "quoteId", "recommendedTopUp", "shortage", "walletBalance" ],
        "type" : "object"
      },
      "Talisman" : {
        "properties" : {
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "csrfHeader" : {
        "description" : "세션 기반 상태 변경 요청에 필요한 CSRF 토큰",
        "in" : "header",
        "name" : "X-CSRF-Token",
        "type" : "apiKey"
      },
      "sessionCookie" : {
        "description" : "카카오 로그인 후 발급되는 HttpOnly 세션 쿠키",
        "in" : "cookie",
        "name" : "JSESSIONID",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "description" : "뿌기사주 백엔드 API 문서",
    "title" : "뿌기사주 API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/quotes/fortune" : {
      "post" : {
        "operationId" : "quote",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseSuneungQuoteResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "suneung-reading-controller" ]
      }
    },
...
    "/api/v1/reading-purchases" : {
      "post" : {
        "operationId" : "purchase",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PurchaseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePurchaseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "suneung-reading-controller" ]
      }
    },
...
    "/api/v1/readings" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "fortuneType",
          "required" : false,
          "schema" : {
            "enum" : [ "OVERALL", "LOVE", "WEALTH", "COMPATIBILITY", "SINSAL", "SUNEUNG" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "personId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReadingListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "suneung-reading-controller" ]
      }
    },
...
    "/api/v1/readings/{readingId}" : {
      "get" : {
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "readingId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReadingResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "suneung-reading-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}